Changes for rename method.
Changes added for rename method, and unit test added for rename
directory with files

+++ src/Utilities/BasicOperation.java
@@ -179,7 +179,10 @@ public class BasicOperation {
      * @return true if the file with the new name exists.
      */
     public static boolean renameItem(File oldItem, File newItem) {
-        boolean result = oldItem.renameTo(newItem);
-        return (result &&(!oldItem.exists()));
+        boolean result = false;
+        if (!newItem.exists())
+            result = oldItem.renameTo(newItem);
+        
+        return (result && (!oldItem.exists()));
     }
 }
